+++ packages/navigation/app-navigation/src/components/IconButton/types.ts
@@ -1,15 +1,17 @@
 import { ButtonProps } from '@atlaskit/button';
+import { AppNavigationTheme } from '../../theme';
 
 export type IconButtonProps = Pick<ButtonProps, 'onClick'> & {
   className?: string;
   icon: ButtonProps['iconBefore'];
   testId?: string;
-  theme?: ButtonProps['theme'];
+  theme: AppNavigationTheme;
   tooltip: string;
 };
 
 export type IconButtonSkeletonProps = {
-  dimension?: number;
+  className?: string;
   marginLeft?: number;
   marginRight?: number;
+  size?: number;
 };