+++ packages/navigation/app-navigation/src/components/PrimaryButton/skeleton.tsx
@@ -1,1 +1,7 @@
-export { PrimaryButtonSkeleton } from './styles';
+/** @jsx jsx */
+import { jsx } from '@emotion/core';
+import { primaryButtonSkeletonCSS } from './styles';
+
+export const PrimaryButtonSkeleton = () => (
+  <div css={primaryButtonSkeletonCSS} />
+);