+++ packages/navigation/app-navigation/src/components/Search/index.tsx
@@ -1,24 +1,26 @@
 /** @jsx jsx */
 import SearchIcon from '@atlaskit/icon/glyph/search';
-import { B50 } from '@atlaskit/theme/colors';
-import { jsx, SerializedStyles } from '@emotion/core';
+import { jsx } from '@emotion/core';
 import { Fragment } from 'react';
 
-import { IconButton } from '../IconButton';
+import { ThemedIconButton } from '../IconButton';
 import { TriggerManager } from '../TriggerManager';
 
-import { SearchInput, SearchWrapper, IconWrapper } from './styled';
-import { searchIconStyles, searchInputStyles } from './styles';
+import {
+  searchInputContainerCSS,
+  searchIconCSS,
+  searchInputCSS,
+  searchInputIconCSS,
+} from './styles';
 import { SearchProps } from './types';
 
 type SearchComponentProps = {
-  css: SerializedStyles;
   onClick: SearchProps['onClick'];
   text: SearchProps['text'];
 };
 
 const SearchComponent = (props: SearchComponentProps) => {
-  const { onClick, text, ...searchProps } = props;
+  const { onClick, text } = props;
 
   const onChange = (...args: any[]) => {
     // @ts-ignore
@@ -26,17 +28,18 @@ const SearchComponent = (props: SearchComponentProps) => {
   };
 
   return (
-    <SearchWrapper {...searchProps}>
-      <IconWrapper>
-        <SearchIcon label={text} primaryColor={B50} />
-      </IconWrapper>
-      <SearchInput
+    <div css={searchInputContainerCSS}>
+      <div css={searchInputIconCSS}>
+        <SearchIcon label={text} />
+      </div>
+      <input
+        css={searchInputCSS}
         placeholder={text}
         onChange={onChange}
         onClick={onClick}
         value=""
       />
-    </SearchWrapper>
+    </div>
   );
 };
 
@@ -47,13 +50,9 @@ export const Search = (props: SearchProps) => {
     <TriggerManager {...triggerManagerProps}>
       {({ onTriggerClick }) => (
         <Fragment>
-          <SearchComponent
-            css={searchInputStyles}
-            onClick={onTriggerClick}
-            text={text}
-          />
-          <IconButton
-            css={searchIconStyles}
+          <SearchComponent onClick={onTriggerClick} text={text} />
+          <ThemedIconButton
+            css={searchIconCSS}
             icon={<SearchIcon label={tooltip} />}
             onClick={onTriggerClick}
             tooltip={tooltip}
@@ -63,5 +62,3 @@ export const Search = (props: SearchProps) => {
     </TriggerManager>
   );
 };
-
-export { SearchSkeleton } from './skeleton';