+++ packages/navigation/atlassian-switcher/src/components/prefetch-trigger.tsx
@@ -1,8 +1,7 @@
 import * as React from 'react';
 import throttle from 'lodash.throttle';
+import { prefetch } from '../prefetch';
 import now from '../utils/performance-now';
-import { prefetchAll } from '../providers/instance-data-providers';
-import { prefetchAvailableProducts } from '../providers/products-data-provider';
 import {
   NAVIGATION_CHANNEL,
   NavigationAnalyticsContext,
@@ -30,6 +29,7 @@ const TRIGGER_CONTEXT = {
 };
 
 type PrefetchTriggerProps = {
+  product?: string;
   children: React.ReactNode;
   cloudId?: string;
   Container?: React.ReactType;
@@ -54,13 +54,8 @@ class PrefetchTrigger extends React.Component<
 
   private triggerPrefetch = throttle(
     () => {
-      const { cloudId } = this.props;
-      if (cloudId) {
-        prefetchAll({ cloudId });
-      }
-      if (this.props.enableUserCentricProducts) {
-        prefetchAvailableProducts();
-      }
+      prefetch(this.props);
+
       this.fireOperationalEvent({
         action: 'triggered',
       });