+++ packages/navigation/atlassian-switcher/src/utils/__tests__/unit/links.tsx
@@ -11,6 +11,7 @@ import {
   ProductLicenseInformation,
   LicenseInformationResponse,
   ProductKey,
+  Product,
 } from '../../../types';
 
 import { resolveRecommendations } from '../../../providers/recommendations';
@@ -164,12 +165,14 @@ describe('utils/links', () => {
   });
 
   describe('getAdministrationLinks', () => {
+    let isEmceeLinkEnabled = false;
     it('should assemble admin links for site admins', () => {
       const isAdmin = true;
       const isDiscoverMoreForEveryoneEnabled = false;
       const result = getAdministrationLinks(
         isAdmin,
         isDiscoverMoreForEveryoneEnabled,
+        isEmceeLinkEnabled,
       );
       const expectedResult = [`/admin/billing/addapplication`, `/admin`];
       expect(result.map(({ href }) => href)).toMatchObject(expectedResult);
@@ -180,6 +183,7 @@ describe('utils/links', () => {
       const result = getAdministrationLinks(
         isAdmin,
         isDiscoverMoreForEveryoneEnabled,
+        isEmceeLinkEnabled,
       );
       const expectedResult = [
         `/trusted-admin/billing/addapplication`,
@@ -192,10 +196,41 @@ describe('utils/links', () => {
       const result = getAdministrationLinks(
         true,
         isDiscoverMoreForEveryoneEnabled,
+        isEmceeLinkEnabled,
       );
 
       const expectedResult = [`administration`];
       expect(result.map(({ key }) => key)).toMatchObject(expectedResult);
+    });
+    it('When product is Jira & Emcee enabled, should include Jira Emcee link', () => {
+      const product = Product.JIRA;
+      isEmceeLinkEnabled = true;
+      const isDiscoverMoreForEveryoneEnabled = false;
+      const result = getAdministrationLinks(
+        true,
+        isDiscoverMoreForEveryoneEnabled,
+        isEmceeLinkEnabled,
+        product,
+      );
+
+      const expectedResult =
+        '/plugins/servlet/ac/com.atlassian.jira.emcee/discover';
+      expect(result.map(({ href }) => href)).toContain(expectedResult);
+    });
+    it('When product is Confluence & Emcee enabled, should include Confluence Emcee link', () => {
+      const product = Product.CONFLUENCE;
+      isEmceeLinkEnabled = true;
+      const isDiscoverMoreForEveryoneEnabled = false;
+      const result = getAdministrationLinks(
+        true,
+        isDiscoverMoreForEveryoneEnabled,
+        isEmceeLinkEnabled,
+        product,
+      );
+
+      const expectedResult =
+        '/wiki/plugins/servlet/ac/com.atlassian.confluence.emcee/discover';
+      expect(result.map(({ href }) => href)).toContain(expectedResult);
     });
   });