+++ packages/core/global-navigation/CHANGELOG.md
@@ -1,5 +1,40 @@
 # @atlaskit/global-navigation
 
+## 8.0.0
+
+### Major Changes
+
+- [major][f0305e1b06](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/f0305e1b06):
+
+  Major breaking change. atlassian-switcher has been unbundled from global-nav. Users of global-nav will have to explicitly pass their own rendering of atlassian-switcher into global-nav via appSwitcherComponent props
+
+## 7.7.1
+
+- Updated dependencies [8d0f37c23e](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/8d0f37c23e):
+  - @atlaskit/drawer@5.0.10
+  - @atlaskit/dropdown-menu@8.1.1
+  - @atlaskit/modal-dialog@10.2.1
+  - @atlaskit/navigation-next@6.6.2
+  - @atlaskit/tag@9.0.7
+  - @atlaskit/avatar@17.0.0
+  - @atlaskit/theme@9.2.2
+
+## 7.7.0
+
+### Minor Changes
+
+- [minor][c68c9bc048](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/c68c9bc048):
+
+  Adds props to allow global navigation to control the focus lock for all the drawers
+
+## 7.6.5
+
+### Patch Changes
+
+- [patch][6410edd029](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/6410edd029):
+
+  Deprecated props, `value` and `onValueUpdated` have been removed from the Badge component. Please use the children prop instead.
+
 ## 7.6.4
 
 - Updated dependencies [a75dfaad67](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/a75dfaad67):